Semi-rationalize get_dummy_expr_field_value implementations (#7314)
If there is an actual type involved that might matter,
get_dummy_expr_field_value ought to produce a value of that type.

Write one helper function for this in `edb.schema.types` and call it
everywhere that cares.

This fixes at least one bug involving globals in a somewhat twisted
interaction with policies (though I am still doing some more thinking
about those cases, so I may fix it in an additional way in a
follow-up).