Fix an UNLESS CONFLICT on links perf regression (#7349)
The regression was introduced in #7284, which made
`utils.subject_substitute` substitute `__subject__` in for the
implicit path prefix. That fixed an issue in some cases, but
introduced an extra level of Set/SelectStmt indirection in
others.

That level of indirection could cause us to emit redundant semijoins
in some cases, which caused a major performance regression on UNLESS
CONFLICT on object level exclusive constraints that referred to a link
using a partial path prefix. (Referring to the link using
`__subject__.` was already slow!)

Fix that by making `subject_substitute` smarter.

The redundant semi-join issue can come up in other cases too, like:
```
select (User, (User,).0.deck);
```