Add test that we only create on index on id (#7414)
This is currently broken due to #7413.

The main bulk of this PR is reintroducing (again, and probably for
good this time) machinery to establish an asyncpg connection to the
backing postgres database. This lets us query pg_catalog to see how
many indexes there are.
We probably want more tests like this.