Fix ai::to_context duplication
The problem is that when we use the new inlined `body` field of
`FunctionCall`, the backend compiler still always compiles the
arguments, and since the arguments weren't registered in the scope
tree, the compiled arguments might not get picked up.

Fix this by always registering arguments when we produce an inlined body.
(We could *not* compile the arguments unconditionally for inlined
functions like this, but we'd still need to register the sets in case
the argument is used twice).

Fixes #7463.