Add an internal-only TEMPLATE BRANCH that directly uses sql templates (#7477)
The test suite used to rely on this for cloning databases, and switching
away caused some big slowdowns. Both because this method is somewhat faster
but more because the new method only allows one branch at a time to avoid
going overbudget on connections.

Progress on #7329.
On my machine this speed up edb test from 13:49 to 11:20.