Fix a race condition where older database configs can overwrite newer ones (#7485)
If two configs are done in short succession, both will fire off tasks
to do `introspect_db`, and it is possible that the first will read the
config after the first change, but not be able to reflect that change
until the second configure has finished, which will cause the older
config to predominate.

This does not actually fix #7330 as described, but it should fix the
current flakes that are still happening.