Fix nested config objects that add new exclusive props
We use the exclusive properties to generate the id of an object, but
this causes trouble is a subtype adds new ones, since we won't take
that into account when generating the ids in the link.
Account for that properly by only using the target type's props.

No test, but an incoming PR from Scott will exercise this.