Different approach to unbreaking cache recompilation after restart
In #7515 I solved it by including schema_version in the serialized
data and bumping the protocol version to indicate this. The server
startup is currently not resilent to seeing new protocol versions,
though, and so it fully breaks rolling back to a new version.

Fix it instead by appending the schema_version *without* bumping the
protocol version.

We should do another follow-up to make cache hydration ignore new
versions.