Create a task for each DB to ensure we generate the correct load (#7524)
The previous QoS tests were "slipping" and generating an insufficient amount of load when the overhead of the pool was too high. This modifies the query generation to create one task per DB and sleep an appropriate amount of time. The tests were unable to hit the query/sec numbers requested and this was causing the QoS numbers to be lower than they should have been. The QoS output for the various tests change to reflect this, but the overall behaviour of the pool is identical.

Note that this had the side-effect of increasing the total number of queries run in most of the tests, which lowering most of the ConnectionOverhead scores. The ConnectionOverhead scoring now uses "# of disconnections / # of queries" which is a more stable number and can be more easily compared across tests.