Clean up the instance metadata setup during bootstrap
Currently, the stdlib SQL we generate has a tenant ID hardcoded in,
but we do some hacky stuff when generating the tpldbdump to patch it
up there.

The upshot is that if you have a cached stdlib but not a cached
tpldbdump, then using alternate tenant ids won't work. Sigh.

This is pretty marginal but my in-place upgrades WIP is disabling
tpldbdump, so I ran into it.

Fix this by computing the DB name dynamically.