Adjust how box2d and box3d are encoded in binary.
The box2d and box3d types don't have a binary encoding in PostGIS, they
are castable into geometry, though. Many of the PostGIS functions seem
to rely on these box types to be implicitly castable to geometry, so we
use that to cast them into geometry in order to transfer over binary
protocol.