Rust connection pool/Python integration (#7544)
Follow-up work from #7478

Note that running EDGEDB_USE_NEW_CONNPOOL=1 edb test fails with a few tests. These will get sorted out as we move forward. The tests unfortunately require the logging to match the expected output from the old pool, which is a small challenge. I'll figure out if it makes sense to rewrite these tests in the next round.

The interaction between Rust and Python is pretty straightforward: we use a cross-thread channel to send messages and we pair that with a self-pipe that we write to to notify the asyncio end that there's a new message. We preemptively read in a few places to reduce latency, and if we manage to pick up a message, reduce the number of messages we expect to read async by one.

As we move forward, these patterns will probably get moved into a more common set of code rather than being adhoc like this.

What's left after this PR:

    Complete garbage collection work -- connections should not get GC'd too quickly
    Export metrics/logs