add a failing test
when I have `INSERT INTO "Document" (title) VALUES ('Report'), ('Briefing')`
The input relation can have an interator, but it does not have an identity & value.
I have cases when using iterator as identity/value will try to do a join with base
table and compare ids.

If I remove the identity/value, the ql compiler will complain becuse it is
first doing a plain compile_Set on the iterator.