Fix TagUnpackRow encoding over SQL adapter (#7686)
A sister PR to #7684

When there is DML without RETURNING clause, we inject a COUNT(*) to the
top-level compiled query. We then intercept the single row produced and
convert it tag of to `CommandComplete` message.

This PR makes sure that `Bind` of such queries requests the result to be
in binary encoding and not the encoding that that client might have
requested. What they have requested does not matter, since we are not
returning any results.

This does also simplify our code for parsing the data row into the tag.