Rust connpool: hook up metrics (#7681)
This is the last part of the connpool work which brings the work up to
parity with the previous connpool. Note we may need to make some
additional tweaks for QoS in the future, as the metrics are somewhat
sensitive to overall timing. I don't think this means our pool is
malfunctioning in any way, just that both the Rust and Python pools are
very sensitive to the order of incoming connections.

There are two failing tests that were disabled, as they have coverage in
the new pool code. One test behaves differently in the new pool by
design.

Some refactoring took place in this PR to clean up dependencies and the
knob optimizer itself.