Support installing packaged extensions
An packaged extension is a .zip file with a MANIFEST.toml file.
Installing an extension consists of:
 1. Dropping the .zip file into the extensions/ search path of edgedb
 2. If present, installing the postgres extension files into
    the packaged postgres directory.

The installation is currently done by an `install-ext.py` python
script, which should get turned into part of the CLI.

On server startup, the server checks the search path for any .zip
files, checks the MANIFEST for the names and versions, and if the
package does not exist, runs the script to install it.

The major rationale of having a separate MANIFEST with name/version is
so that we don't need to parse the whole extension package in order to
determine the name and version of extension, but maybe that isn't
necessary. We could just parse to do it, or we could try getting those
from the filename?