Fix ast transformer not working with sets of tuples of nodes. (#7760)
Some fields are sets of tuples. The current visit implementation will produce attempt to produce a set of lists, however, lists are not hashable. This causes `NodeTransformer.generic_visit` visit to fail to create the new field data.

The return type should not be modified in `NodeVisitor.container_visit` since most visitors are not expected to return the exact type of the node when visiting. It is possible to override the function in `NodeTransformer`, but this change is minimal.