Fix ISE involving UNLESS CONFLICT and WITH interaction
We are directly invoking `infer_volatility` on a Pointer in some
cases, and that isn't supported yet. Fix it in the simplest way by
making the logic behave like it did before the Pointer refactor.

It is actually also easy to fix the `infer_volatility` shortcoming
(and I will), but if we just do that, we run into a spurious
volatility error that I need to track down.

Fixes #7433.