Reduce the postgres INFO spew when using function cache
Whenever postgres skips a DROP because IF EXISTS fails, it emits an
INFO. Since this happens every time a query cache misses, I found this
annoyingly noisy.

Make those operate by using an explicit condition check instead. Doing
this required removing a check that converted such checks into IF
EXISTS as well as fixing some problems with our checks.

This is one of two checkboxes left on #6493, I think, the other being
a flake I dimly remembered, so we can probably close that?