Fix a regression involving optional arguments and assert_exists (#7798)
A regression was introduced in #7763 having to do with determining
whether paths are optional in the scopetree. Most of the time, having
spuriously optional paths is just a performance issue, but in some
cases (like the assert_exists calls injecting into `required link`s
when there is an access policy on the target), having a path be
optional can result in an assert_exists error firing when really the
assert_exists shouldn't have been evaluated at all.

The cause of the regression was that the computation of whether an
existing node that is being factored is underneath a fence was wrong,
and that wrong computation was being used in more cases now.