Initial support for SQL over native protocol
The binary protocol version is bumped to 3.0 to accommodate the addition
of the new `input_language` field to the `Parse` and `Execute` messages.
The field can be set to `b'E'` for EdgeQL or to `b'S'` for SQL.  The
result of a query sent as SQL is always encoded as a free object.  In
other regards the SQL execution and caching pipeline is identical to
EdgeQL.

Caveats and limitations:

Not all query features are supported when in SQL mode, specifically
cardinality assertions and output format are always `MANY` and `BINARY`
correspondingly.

DDL and explicit prepared statements are unsupported.

Multi-statement queries are not supported yet.