Rust PostgreSQL connection client (#7605)
Implements a basic Rust postgres connection library and re-uses its
internal state machine to replace the guts of `pgcon`.

User-facing changes:
- Bugfix: EdgeDB will choose the smaller of server `max_connections` or
the maximum number of connections defined for the user's current `role`
to prevent hitting connection limits.
- Change: `sslmode`: SSL is always requested unless sslmode=disable. If
the server returns `S` to indicate SSL support and then fails to accept
the user credentials, we do not retry without SSL.
- Feature: Plain-text password authentication in backends is now
supported, though discouraged.

Changes:

- New `PGRawConn` that behaves as a transport and fully connects to
postgres before returning the transport to the protocol. This transport
can then be used transparently and with zero overhead as it only uses
Rust code for connection setup.
- New `source_description` parameter wherever we use connections
directly or indirectly to assist with tracking down what particular part
of the system requested a connection. This is not currently propagated
into the parameters we send to the server, but we could do this in the
future.
- Major improvements to DSN parsing, as well as moving all DSN code to
Rust. There is now a thin shim in `pgconnparams.py`, but it delegates
all logic to the Rust side of things.
- `class CreateParamsKwargs(TypedDict)` is used wherever we hold a bag
of connection parameters to assist with typing.
- `pgaddr` has been removed from the EdgeDB server settings as it was
not sufficient to connect to the backend. It has been replaced with a
fully-rendered DSN. `edgedb-rust` will need to be updated to support
this.

## Testing strategies

DSN parsing: we have a full battery of DSN parsing tests that are
compared against libpq. The tests that were previously implemented for
pgconnparams have been migrated into this rust codebase. The tests used
internally for postgres to test DSN parsing were also extracted and
added to the library. On top of this, a number of "hard" examples were
added to exercise the code.

Auth code: auth code is tested against a "capture" postgres that we
inittb, boot up on a socket and test. Tests for each auth method (scram,
md5, password) are tested with and without SSL.

## DSN notes

The DSN parsing code behaves effectively the same as the previous
version of the code with a few small differences:

- Backslash escaping/unescaping in pgpass and the connection string now
fully match the behaviour of libpq
- A few additional warnings were added in cases where an explicit
passfile was specified but cannot be read (we still don't warn for
implicit passfiles).

## Examples:

Parse a DSN:
```
cargo run --example dsn 'postgres://username@/database_name?host=path-to-socket'
```

Query a running server:
```
cargo run --example connect -- -U foo -P secret -u <<direct path to unix socket .s.PGSQL.5432>>
```

---------

Co-authored-by: Fantix King <fantix.king@gmail.com>