Implement warnings in the server (#7823)
Warnings are encoded in JSON as a header for
CommandDataDescription. Users of the clients should be able to provide
a callback that is invoked on the list of warnings; the default will
be to log them.

Adds a test function, `std::_warn_on_call()` that generates a warning
whenever we call it.

The most annoying part about all of this is collecting warnings
generated when compiling expressions in the schema and DDL.

Implements the server component of #7822.