Forbid certain system functions over SQL adapter (#7829)
PostgreSQL supports certain SQL functions that allow [access to the
filesystem](https://www.postgresql.org/docs/16/functions-admin.html#FUNCTIONS-ADMIN-GENFILE),
managing backups, WAL and other administration functions of the
database.

This is not intended behavior of EdgeDB and goes against security model
of EdgeDB. To our knowledge, this feature is also not widely used by
users of PostgreSQL. Because of that, we are forbidding a range of
PostgreSQL system administration functions.

This change only affects access to EdgeDB via SQL adapter. Connections
directly to underlying PostgreSQL instance are not affected.