Pass in cardinality_mod of json_parameters
`<required ..>$xx` had been evaluated as AT_MOST_ONE with json_parameters on,
even though it has a runtime check that guarantees non-empty values. This issue
was captured with modifying function `std::net::schedule_request()` that errors
if the required parameter value doesn't have cardinality ONE.

The issue seems to be a missing cardinality_mod, which is used to detect if a
JSON typecasting should be widen to include empty values.