Remove extraneous edgedb-session
The only place where we should be sending a session token is in the
`/token` PKCE exchange endpoint. I believe these bits were leftover from
the pre-PKCE implementation, and in any case are the "wrong" place for
the token cookie which should be between the end-user's user
agent (browser) and the application's web server.