Initial implementation of "RFC 1027 - Simplfying path resolution" (#7809)
See [RFC 1027 - Simplifying path
resolution](https://github.com/edgedb/rfcs/blob/master/text/1027-no-factoring.rst)
and #7766.

There are still some improvements I would like to make to the
`warn_old_scoping` mode, and we should update and triage more of the
test suite to use `simple_scoping` and/or `warn_old_scoping`.
    
Note that while a lot of tests rely on path factoring, I expect this is
very unrepresentative of actual EdgeQL in the wild. There are a lot of
tests that use path factoring because path factoring needed a lot of
tests to make sure it was right.