Add the ability to run internal queries with reduced tx isolation
Internal queues, such as `net::http::ScheduledRequest` are structured
in a way that produces lots of serialization errors if `serializable`
isolation is used.  To side-step this, add a way to run internal queries
with reduced isolation level (via the new `tx_isolation` argument to
`execute()` and `execute_json()`.

As a proof of concept, switch `net_worker` to `repeatable read`.