Make sure that simple inlined INSERTs don't select from the table
The code we were generating for inlined calls to a function containing
an INSERT was selecting from the table (overlayed with the INSERT)
after performing the insert.

This is not incorrect, exactly, but it's suboptimal, and we think that
this might be the source of the mysterious transaction serialization
errors that we have been seeing in the std::net tests.

The source of the issue is that the function call was only producing a
`value` aspect, but to select properties out of it a `source` aspect
was needed, and so `ensure_source_rvar` joins one back in.

Fix it by making sure we can output a source aspect from an inlined
function call.