Try to fix some of the DROP DATABASE flakes (#7907)
I think one of the sources of trouble might be that _early_introspection
is hanging on to connections and can't be forced to drop them.
Try to work around it by forcing them closed at the end of introspecting.

Also, when in dev/test mode, dump out all the info about the other
connections.  Hopefully this will let us figure out where they are
coming from.