Implement an --uninstall command for load_ext
Note, though, that this only uninstalls the extension from the "server
installation".

The extension package may have already been loaded into the
sys::ExtensionPackage tables of instances, though, and this is not
able to remove it from there.

Removing a package that is actually in use will probably cause
trouble.