Fix embedding data being stored with the wrong entry (#7932)
The batching of inputs based on token limits implemented in #7574 introduced a bug where the embeddings ids would not be reliably paired with the embeddings. This causes the wrong results to be returned in ai searches.

This error was not caught in the existing tests because the `custom::test` provider used in tests does not have a tokenizer and so it doesn't use the batching. The last commit adds a tokenizer, and the added test will fail if cherry picked to master.