Pass through source OIDC id_token
We currently only use a small amount of data from the identity
provider's ID token, but applications might want to use different bits
of this data for their own user profile information. We currently
require them to fetch this on their own, but this will make this process
slightly easier by returning the ID token, if we have one.