Fix an error where schema_links wasn't properly namespaced (#7953)
The issue is we were running `trampoline.fixup_query` *before*
substituting in the schema name.

This probably *shouldn't* show up in any real situations, since
it is only used to produce an error message, and we shouldn't
actually be hitting link policy error messages for the schema.

No real way to test this as a result. I tested it against a bug I had
introduced while working on something else.