Remove special DeltaRoot.apply_subcommands function (#7955)
It was overloaded to run subcommands in a certain order, but it
doesn't seem to be necessary anymore. (And I can't think of why it
would be as things are.)

Dropping it fixes a bug where a 'caused' subcommand would get run
twice, which in practice could cause some mismatches in whether a
collection is deleted in the `_redo_everything` path of pgsql/delta.