Add a debug flag to validate schema reflection (#7958)
When the flag is set, after every DDL command we reintrospect the
database and compare that to the in-meory schema. I've recently landed
a few PRs fixing issues exposed by this. (IIRC, I also fixed a few
earlier this year when I first wrote this PR.) Overall the system is
pretty solid.

Run it in CI once a day, since I think it is too slow to run all the
time.

Fixes #5169.