Allow OAuth callback to be specified directly (#8022)
Currently, we expect that the OAuth Identity Provider should always
redirect back to the auth extension's server endpoint, so we build this
URL ourselves. However, there might be times when users want to control
the OAuth flow themselves, so the callback should redirect to some URL
that they've specified.

One confusing thing here is that sometimes the URL provided at
`redirect_to` might be refered to as the "callback" URL, but that URL
specifies where the auth extension server should call the application at
the end of the flow.