Fix email button background (#7974)
When passing `None` explicitly, which we are always doing, the fallback
color was not being set. It was also the wrong format considering that
we store them without the leading `#`.

---

I tried for an unreasonable amount of time to add tests for this, but
the configuration was not stable when running the full test suite, so I
just reverted back to no additional tests.