Replace headers with annotations in Parse/Execute (#8037)
Take annotations (`dict[str, str]`) instead of headers (`dict[int,
bytes]`) in Parse/Execute messages.

We forgot this in protocol 1.0 but had been ignoring headers, so it was
fine until we're trying to make use of annotations now. This change is a
part of the binary protocol version 2.0 (EdgeDB 6.0a1 or nightlies
before this PR will not work properly with annotations).