Fix bit string literals over SQL adapter (#8112)
So, bytea and bitstring/bit varying are not the same thing in pg. In
SQL, they are both represented by string literals (of type unknown). The
difference is that bitstring can have sub-byte length.

Initially, I represented BitStringConstant with bytes + bitlength. It
turned out that Python has no easy way to convert 10101010001 to bytes,
so I replaced that with plain str, as it was passed from parser.

We do still need `ByteaConstant`, because in EdgeQL's bytes literal
should land in bytea pg type, not bitstring.

---

Here is the now unused code for codegen of the old BitStringConstants:

```
def quote_bitstring_literal(data: bytes, length: Optional[int] = None) -> str:
    """Return valid SQL representation of a bit string value."""

    if data:
        
        b = binascii.b2a_hex(data).decode('ascii')
        if length:
            b = b[0:((length + 3) // 4)]
            if length % 4 != 0:
                return f"x'{b}'::bit({length})"
        return f"x'{b}'"
    else:
        return "x''"
```