Implement position mapping for SQL over native proto
This is a little fiddly, since we need to account for query
normalization. We do this by traversing the pre and post normalization
ASTs in parallel, building up a map.  In my testing, postgres only
returned error positions for errors that could be detected while
describing the query, so to simplify the implementation we only remap
errors that happen there.

Also do a better job translating the errors:
 * Don't ISE when it doesn't map to an edgedb error
 * Preserve hint and details when we can

Fixes #8077.