Cache normalized queries over SQL adapter (#8086)
Enables query normalization for sql-over-pg. This uncovered a bunch of
bugs that were also present on already enabled sql-over-edgedb-proto,
but have not been caught because we didn't have many tests.

This PR also contains two commits to libpg_query. I suggest reviewing
commit by commit.

Encountered problems:
- `select 1 + '1'` should infer that `'1'` has `integer` type. The
solution was that string constants need to be reported to postgres as
having type `unknown`.
- `SET TRANSACTION ISOLATION LEVEL read uncommitted` was being
normalized to `SET TRANSACTION ISOLATION LEVEL $1`, which does not even
parse. Solution was to patch libpg_query to only normalize SELECT,
UPDATE, DELETE, INSERT and MERGE.
- `SELECT NULL, 42` was being normalized to `SELECT NULL, $2`. A bug in
libpg_query (that we introduced, not the original author),
- `SELECT current_schemas(true)` was normalized to `SELECT
current_schemas($1)`, which meant that we are unable to statically
evaluate the query. Solution was `DisableNormalization` exception that
re-compiles the query without normalization and disables caching the result.
- `SELECT pg_get_serial_sequence('a', 'b')` was normalized to `SELECT
pg_get_serial_sequence($1, $2)` and resolved to `SELECT NULL`, which
tripped PostgreSQL, because it was unable to determine type of the two
params, which we specified as `unknown`. The solution was to detect that
the params are unused and report them as being `text`.
- `SELECT ROW('hello')` is normalized to `SELECT ROW($1)`, which tripped
PostgreSQL, because it was unable to determine type of the `unknown`
param. Solution was to compile fields in `ROW` with additional type cast
to `text`.
 

---------

Co-authored-by: Michael J. Sullivan <sully@msully.net>