Send `identity_id` on require_verification (#8129)
During the sign up flow, we want to allow applications to still create
and associate an Identity, since further flows like the verification
flow might happen as a sign in, such as when the user verifies their
email from a different device than they signed up with. That changes the
semantics a bit from "we hide the identity from the application until
the email is verified" to "we do not give the application an auth token
until the email is verified" which seems more honest.