Change some pg_catalog views to be MATERIALIZED (#8145)
This *hugely* speeds up most pg_catalog SQL introspection operations,
like
`\d` and pg_dump.

Unfortunately it slows down DDL operations, because we need to refresh
the
views. On my machine, it looks like it typically adds about 150ms to
applying a migration.

This is probably fine for our users, but is pretty bad for us because
our
test suite is pretty heavy on DDL; this slows it down on my machine from
15.5m to 19.5m, which might be too much to merge.
Thoughts?

Fixes #8137.