Try adding a fence between pipe and channel operations in http library
This is an experiment that I don't expect to work, aimed at fixing
flakes like https://github.com/edgedb/edgedb/actions/runs/12662375681/job/35288956004
where something is hanging in the rust/python async HTTP library.

The (unlikely seeming to me) hypothesis here is that the pipe
write+read is not establishing a "synchronizes-with" relationship
and so is missing the message to the queue.

I don't really think this is possible, though, and even if it were I
kind of expect that it would still probably work.