Use rustls for http fetches (#8201)
@msullivan and I tracked down an issue where HTTP tests were causing
crashes on aarch64 runners. The culprit turned out to be `openssl-probe`
calling `setenv()` in a multithreaded environment.

Switch http to rustls + native certs. We still have the probe dep in
Cargo.lock, but as long as we don't call the functions that set envvars,
we should be fine.

This was the original crashing stack trace -- basically something in
asyncio was trying to format an error (illegal seek, we believe) at the
same time we were initializing the HTTP code `openssl-probe` was
attempting to set some OpenSSL envvars.

```
#0  0x0000ffff805a3e90 in getenv () from /chroot/lib64/libc.so.6
#1  0x0000ffff8059c174 in __dcigettext () from /chroot/lib64/libc.so.6
#2  0x0000ffff805f263c in strerror_r () from /chroot/lib64/libc.so.6
#3  0x0000ffff805f254c in strerror () from /chroot/lib64/libc.so.6
#4  0x00000000005bb76c in PyErr_SetFromErrnoWithFilenameObjects ()
#5  0x00000000004e4c14 in ?? ()
#6  0x000000000049f66c in PyObject_VectorcallMethod ()
#7  0x00000000005d21e4 in ?? ()
#8  0x00000000005d213c in ?? ()
#9  0x00000000005d1ed4 in ?? ()
#10 0x00000000004985ec in _PyObject_MakeTpCall ()
#11 0x00000000004a7734 in _PyEval_EvalFrameDefault ()
#12 0x000000000049ccb4 in _PyObject_FastCallDictTstate ()
```