Implement \(expr)-style string interpolation (#8210)
`\(` is currently rejected inside string literals, so there is no
compatability issue.

Expressions in the `\(expr)` blocks get automatically cast to str.

The implement is based on the lexer producing `STRINTERPSTART`,
`STRINTERPCONT`, and `STRINTERPEND` tokens that are parsed by the
parser.  The lexer needs to be a little stateful, unfortunately, to
track what kind of opening quote opened the string and to track
the nesting level of parenthesis.

Fixes #272.