Monitor open FDs (#8217)
Add an FD monitor for servers. This is potentially something we can do
more efficiently in Rust at a later date, but a Python implementation
will allow for an easier backport to 5.6.

We use a new shortcut available in modern Linux kernels to avoid
counting directory entries (os.stat's size == # of open FDs).

Threads are used to simplify the implementation.