[5.x] Fix dump and MIGRATION REWRITE when there are many (>1000) migrations (#8240) (#8241)
Currently those paths rely on topological sorting of the migration
history.  Our topo sort has various bells and whistles that would make
it irritating to rewrite it non-recursively (not impossible, of
course; there is a theorem to that effect); instead, order the
migrations in a more obvious way.

It takes about a minute to set up that many migrations, so I haven't
added a test. If we think it is worth it I can, though.