Use better histrogram buckets
Use a method of generating buckets based on specifying how many
buckets to use per order of magnitude.

For bytes, I did 4 buckets per binary order of magnitude, because
binary felt sensible to me. Is that stupid, and I should just use
decimal orders of magnitude?

For our bytes field, we use:
(8, 12.0, 16, 24.0, 32, 48.0, 64, 96.0, 128, 192.0, 256, 384.0, 512, 768.0, 1024, 1536.0, 2048, 3072.0, 4096, 6144.0, 8192, 12288.0, 16384, 24576.0, 32768, 49152.0, 65536, 98304.0, 131072, 196608.0, 262144, 393216.0, 524288, 786432.0, 1048576)

And for querys per connection, we use:
(1, 2.0, 4.0, 6.0, 8.0, 10.0, 20.0, 40.0, 60.0, 80.0, 100.0, 200.0, 400.0, 600.0, 800.0, 1000.0, 2000.0, 4000.0, 6000.0, 8000.0, 10000.0)

Fixes #8245.